+++ cypress/locators/radioButton/locators.js
@@ -3,3 +3,4 @@ export const RADIO_BUTTON_COMPONENT = '[data-component="radio-button"]';
 export const RADIO_BUTTON_GROUP_COMPONENT = '[data-component="radio-button-group"]';
 export const RADIO_BUTTON_FIELDSET_COMPONENT = '[data-component="fieldset-style"]';
 export const RADIO_BUTTON_LEGEND_COMPONENT = '[data-component="legend-style"]';
+export const RADIO_GROUP = '[data-component="radiogroup"]';