+++ cypress/support/step-definitions/create-steps.js
@@ -9,7 +9,7 @@ Then('create className on preview is {string}', (classNameParameter) => {
 });
 
 Then('Create component has proper inner color {string} and background-color {string} and border color {string}', (color, backgroundColor, borderColor) => {
-  link().children().should('have.css', 'background-color', backgroundColor)
+  link().find('a').should('have.css', 'background-color', backgroundColor)
     .and('have.css', 'border', `1px dashed ${borderColor}`)
     .and('have.css', 'padding', '12px 12px 10px')
     .and('have.css', 'text-align', 'center')
@@ -18,14 +18,6 @@ Then('Create component has proper inner color {string} and background-color {str
     .should('have.css', 'color', color);
 });
 
-Then('Create element has {word} border on focus', (borderColor) => {
-  switch (borderColor) {
-    case 'golden':
-      link().children().should('have.css', 'outline', 'rgb(255, 181, 0) solid 3px');
-      break;
-    case 'blue':
-      link().children().should('have.css', 'outline', 'rgba(0, 103, 244, 0.247) auto 5px');
-      break;
-    default: throw new Error('There are only two possible focus outline colors');
-  }
+Then('Create element has golden border on focus', () => {
+  link().children().should('have.css', 'outline-color', 'rgb(255, 181, 0)');   
 });