+++ src/components/toast/documentation/info.js
@@ -6,11 +6,11 @@ const info = (
     <p>Toast</p>
     <StoryHeader>Implementation</StoryHeader>
     <p>Import the component</p>
-    <StoryCode>import Toast from {"'carbon-react/lib/components/toast'"}</StoryCode>
+    <StoryCode>import Toast from &quot;carbon-react/lib/components/toast&quot;</StoryCode>
     <p> Additionally you can pass optional props to the Toast component</p>
 
-    <p><StoryCode padded> {'as'}</StoryCode> Customizes the appearence of the toast changing the colour
-    (see the {"'iconColorSets'"} for possible values).
+    <p><StoryCode padded> as</StoryCode> Customizes the appearance of the toast changing the colour
+      (see the &quot;iconColorSets&quot; for possible values).
     </p>
 
   </div>